refactor: extract requireFlagValue helper and rename cross-origin util
Follow-up cleanup on #311:

- Extract requireFlagValue() from resolveHost() in src/config/env.ts. The
  CLI value-less-flag rejection (bare --host, --host=, whitespace-only) and
  its error message were defined inline three times; they now live in one
  reusable helper so the next string-valued flag can reuse the argv-scan
  workaround instead of re-implementing it. parseCommandLineArgs() is left
  untouched — its "true" sentinel is load-bearing for boolean flags.
- Rename src/utils/dns-rebinding.ts -> src/utils/cross-origin.ts (and its
  test) to match what the code actually is: a cross-origin fetch guard, not
  a full DNS-rebinding defense, as the doc comments already state.

No behavior change. env + cross-origin unit tests pass (62).
